Remove schema from api
